+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/provdocument/domain/ProvDocument.java
@@ -1,4 +1,4 @@
-package com.multi.mlpenterpriseapprovalsystem.prov_document.domain;
+package com.multi.mlpenterpriseapprovalsystem.provdocument.domain;
 
 import com.multi.mlpenterpriseapprovalsystem.common.domain.BaseEntity;
 import com.multi.mlpenterpriseapprovalsystem.company.domain.Company;
