+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/provdocument/dto/ReqProvDocumentCreateDto.java
@@ -1,4 +1,4 @@
-package com.multi.mlpenterpriseapprovalsystem.prov_document.dto;
+package com.multi.mlpenterpriseapprovalsystem.provdocument.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
