+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/provdocument/dto/ReqProvDocumentUpdateDto.java
@@ -1,4 +1,4 @@
-package com.multi.mlpenterpriseapprovalsystem.prov_document.dto;
+package com.multi.mlpenterpriseapprovalsystem.provdocument.dto;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
